This commit primarily addressed the following requirements: (1) Your program will output one of the digits 1...6 as its move, indicating that it is choosing the corresponding pit 1...6 to sow from. If it cannot move, output a 0. (2) Your program should display a reasonable representation of the Kalah board after each of its moves and each of its opponent's moves.
Cleaned up the implementation of the no-move ("If a player who has the move has no stones available, the game ends immediately with all the opponent's stones going into the opponent's Kalah") situation. We made move(pit=0) a legal move when there are no other moves available, and I process this move in apply_move by emptying the opponent's pits into the opponent's kalah (which means the board is empty and the game will be over). We also fixed the major bug where we prematurely exited if *either* of the sides had no stones (we didn't wait for the player's turn or for the player to tell us they have no moves available).

Modified the architecture slightly to pass in a format (FMT_FILE, FMT_CMDLN = file or command-line) to instruct the Referee how to print the output.

The no-move situation is now handled gracefully as part of standard processing (next_move/apply_move). Referee output is now as spec'd for FMT_FILE and FMT_CMDLN.

I also fixed a small bug where InteractPlayer did not display error or re-prompt when receiving non-integer input.